requirements.txt
python-telegram-bot==13.15
google-api-python-client
python-docx
Flask
